Restructuring, removed nufft_spread
